MSPSDS-570: Refactor reporter and question flow to commonize their methods and logic
